// packages/contracts/src/generated/api.d.ts
export interface components {
    schemas: {
// ...
        FlowResult: {
            result: null | components["schemas"]["JsonElement"];
        };
// ...
        KioskAccessoryResponse: {
            /** Format: uuid */
            id: string;
            name: string;
            /** Format: int64 */
            price: number | string;
            /** Format: int32 */
            stockQuantity: number | string;
            shopName: string;
        };
        KioskCatalogResponse: {
            /** Format: uuid */
            bouquetId: string;
            name: string;
            /** Format: int64 */
            price: number | string;
            slotCode: string;
            shopName: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}